xulrunner/mozilla/content/xbl/test/test_bug481558.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=481558
-->
<head>
  <title>Test for Bug 481558</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
  <link rel="stylesheet" type="text/css" href="http://example.com/tests/content/xbl/test/file_bug481558.css">
</head>
<body onload="runTest();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=481558">Mozilla Bug 481558</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
function runTest() {
  is ($('display').xblBoundProperty, undefined, "XBL shouldn't be attached");
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
